m_list: fail on invalid parameters
Commands such as "LIST channel" would formerly be interpreted
as an ELIST command without valid filters, i.e. the entire channel list
would be returned. This is usually not the desired behaviour.

Instead, return a start-of-list numeric, followed by a notice informing
the user their parameters were invalid and an end-of-list numeric.
This sort of mirrors the behaviour exhibited by an attempt to
/list #channel, where #channel does not exist.

Fixes #132.