Add basic implementation of need_auth quarantine
This adds a new flag, need_auth, working similar to need_sasl; however,
instead of rejecting unauthenticated connections entirely, they are
allowed to connect but cannot meaningfully interact with users/channels
other than services or +T clients (see below) until they log in.

This is still a rough implementation that only touches core and a core
module. Specifically, unauthenticated users with need_auth cannot:

- join channels
- /knock channels
- be invited to channels
- message channels
- message users other than services
- receive messages from non-service users

Currently, no restrictions beyond these are applied. This might result
in odd behaviour if need_auth users log out while already joined to
channels. Ideally, services should disallow that.

Additionally, the need_sasl flag causes the above as well as its usual
effect. This will affect users attempting to log out after connecting.

Details on modes / numerics touched:

User mode +V is added (only settable on burst) to implement this.
This allows the need_auth status to be broadcast without requiring any
protocol adjustments, so e.g. services may make use of it.

Numeric 719 is added for being unable to send to +V users.
This appears to be unused so far, and is somewhat close to the +g
numerics which aren't a perfect fit but it could be worse.
(ERR_NOSUCHNICK with a custom message could be acceptable, but at least
irssi has custom handling for this, so users might not see the message.)

User mode +T is added to be able to talk to +V users normally.
This is intended for bots that automatically message users on
connection, and should usually be made an oper-only umode via
configuration.
Services (and server notices) do not require umode +T.